Reformat all files

+++ keymaerax-core/src/main/scala/edu/cmu/cs/ls/keymaerax/KeYmaeraXStartup.scala
@@ -1,27 +1,29 @@
+/*
+ * Copyright (c) Carnegie Mellon University, Karlsruhe Institute of Technology.
+ * See LICENSE.txt for the conditions of this license.
+ */
+
 package edu.cmu.cs.ls.keymaerax
 
 import edu.cmu.cs.ls.keymaerax.bellerophon.ExhaustiveSequentialInterpreter
 import edu.cmu.cs.ls.keymaerax.lemma.LemmaDBFactory
 import edu.cmu.cs.ls.keymaerax.tools.KeYmaeraXTool
 
-/**
-  * Startup support functionality.
-  */
+/** Startup support functionality. */
 object KeYmaeraXStartup {
 
   /** Initializes and updates the lemma cache. */
-  def initLemmaCache(logger: (String, Throwable) => Unit =  (msg: String, ex: Throwable) => {
+  def initLemmaCache(logger: (String, Throwable) => Unit = (msg: String, ex: Throwable) => {
     ex.printStackTrace()
     println(msg)
   }): Unit = {
     try {
-      //Delete the lemma database if KeYmaera X has been updated since the last time the database was populated.
+      // Delete the lemma database if KeYmaera X has been updated since the last time the database was populated.
       val cacheVersion = LemmaDBFactory.lemmaDB.version()
-      if (Version(cacheVersion) < Version(edu.cmu.cs.ls.keymaerax.core.VERSION))
-        LemmaDBFactory.lemmaDB.deleteDatabase()
+      if (Version(cacheVersion) < Version(edu.cmu.cs.ls.keymaerax.core.VERSION)) LemmaDBFactory.lemmaDB.deleteDatabase()
       KeYmaeraXTool.init(Map(
         KeYmaeraXTool.INIT_DERIVATION_INFO_REGISTRY -> "true",
-        KeYmaeraXTool.INTERPRETER -> ExhaustiveSequentialInterpreter.getClass.getSimpleName
+        KeYmaeraXTool.INTERPRETER -> ExhaustiveSequentialInterpreter.getClass.getSimpleName,
       ))
     } catch {
       case e: Exception =>
